# docs/swagger.yaml
basePath: /
definitions:
  gin-user-api_internal_model.User:
    properties:
      email:
        description: 'Example: alice@example.com'
        example: alice@example.com
        type: string
      id:
        description: 'Example: 1'
        example: 1
        type: integer
      name:
        description: 'Example: Alice'
        example: Alice
        type: string
    type: object
  gin-user-api_internal_model.UserPatch:
    properties:
      email:
        example: alice.new@example.com
        type: string
      name:
        example: Alice Updated
        type: string
    type: object
  gin-user-api_internal_model.UsersListResponse:
    properties:
      data:
        description: data
        items:
          $ref: '#/definitions/gin-user-api_internal_model.User'
        type: array
      limit:
        description: page size
        example: 20
        type: integer
      next:
        description: next page url
        example: /users?page=2&limit=20
        type: string
      page:
        description: page number
        example: 1
        type: integer
      prev:
        description: prev page url
        example: /users?page=1&limit=20
        type: string
      total:
        description: total items
        example: 100
        type: integer
      total_pages:
        description: total pages
        example: 5
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Simple user API with GORM + MySQL
  title: Gin User API
  version: "1.0"
paths:
  /users:
    get:
      consumes:
      - application/json
      description: List users with pagination
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin-user-api_internal_model.UsersListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/gin-user-api_internal_model.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin-user-api_internal_model.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create user
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Retrieve a user by its ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin-user-api_internal_model.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user by ID
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Partially update a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/gin-user-api_internal_model.UserPatch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin-user-api_internal_model.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Patch user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Replace a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/gin-user-api_internal_model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin-user-api_internal_model.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update user
      tags:
      - users
swagger: "2.0"